Organize as_search methods into AsSearch module
