+++ lib/elastic_record.rb
@@ -3,6 +3,7 @@ require 'active_support/concern'
 require 'active_model'
 
 module ElasticRecord
+  autoload :AsSearch, 'elastic_record/as_search'
   autoload :Callbacks, 'elastic_record/callbacks'
   autoload :Config, 'elastic_record/config'
   autoload :Connection, 'elastic_record/connection'
