+++ lib/elastic_record/model.rb
@@ -5,6 +5,7 @@ module ElasticRecord
         extend Searching
         extend ClassMethods
         include Callbacks
+        include AsSearch
 
         class_attribute :elastic_connection
         self.elastic_connection = ElasticRecord::Connection.new(ElasticRecord::Config.servers, ElasticRecord::Config.connection_options)
